src/library/requirements.txt
jax==0.2.10
numpy==1.21.4
pillow==8.4.0
